Finished testing the PlayerShip controller, and added new test scene
Implemented and tested movement, and attacking functions within the new ArcadeGame1Test scene. Feel free to check that out. Started laying the framework for bullet objects to reference their shooter. This will allow them to differentiate between enemies and increment the points of the shooter upon a successful hit with an enemy. Also added a few testing materials and prefabs for the new test scene.