Added and tested basic implementation of the generic InteractionController
Added InteractionController, which allows one to define the specific rules for an interaction through the event OnInteract, which has default behaviour to simply return to the requesting gameObject a UI Canvas for a Dialog menu with the object. PlayerController also now fully implements the prioritization of interactable objects, and can decide between interactions when multiple colliders are present.