Added GameCabinetController and finished camera transition for interactions
GameCabinet controller for now will simply store a reference to a scene of a subgame, and provide a Play() function to start the scene from the dialogcanvas returned from the InteractController. The main camera now has two states: interacting, and standard. When the player is interacting with an object, all movement will be disabled for the player, and the camera will lerp down to the over the shoulder interact position. When the player hits the interact key again, all of this goes in reverse. In the future, if we would like the camera to never be obstructed in interact view by the players head, we could simply add a static offset to the main camera using a parent gameobject and set its interacting position to the player themselves.