Added functioning UI to GameObject Interface and tidied up
Added a new concpetual interface that could allow UIs to interact ubiquitously with any interactable game object. If this proves to be two limiting, we could simply derive others from InteractControllerUI such as GameControllerUI and it would have Play(), PlayMultiplayer(), Exit(), etc., to make it more specific, and therefore easier to interface with at the expense of cluttering up the interaction system.