Added Respawning for PlayerShip, as well as Spawn Protection
Now when a player dies, there ship will be moved off the player area, but not deactivated (so the script will still run). After 3 seconds, the player will be moved back to the spawn area with 3 seconds of invulnerability. After which they may be killed again. When a player dies, the gameObject is deactivated, and thus removed from the scene. I also reserved a part in the script to record the score.