Implemented basic enemy behaviours and weapons for first iteration
Finished the two enemy weapons in quesiton for this iteration. The laser and standard gun on assaulters and strikers respectivly. Implemented all level one's enemies behaviours based on the docs. May need to add start up behaviours in the future for entering the screen. Hard coded the behaviour for Cancer, as it is much different and unique to it. Made it so it seeks out the player consistantly, and every second it toggles whether or not its arms are open.