Added some more features to Cancer
Although I'm not assigned for this kind of thing I just wanted to add a few more things to the prefabs so we could at least see them work. I also turned many of the spawns so the enemies will walk the correct direction. See last commit for all changes