Improved Cancer, Ranger, and Assaulter behaviors, added IsActive, and tidied up
Made it so Cancer will approach the players until it is at the world origin on the x axis, and the it will occilate and shoot the players with its new pretty blue laser. Made it so ranger and assaulter will slowly appraoch the player even while doing there other behaviors. Added IsActive that when set to false, the enemy will be unable to move or shoot anything.