Added ADDBConnection Skeleton, and Modified ItemMap
Added a commented out version of ADDBConnection until I figure out how to modify the target framework to .NET 4.6.1, and add a reference to System.Data.SqlClient (even after adding it with the NuGET browser, I can't access it).
Added a GetPermission Function to ItemMap that returns whether or not based on the users permission database hash there aloud to equip an item, or display it as unlocked in the shop.