Updated ItemMap to map up to 64 items from the previous 32
Made ItemMap able to store up to 64 items, and 4 types of item. The number of types may techincally vary, but the total number of items able to be stored will be 64 instead of the old 32.  Also added the skeleton for a few helpful functions, and just cleaned up around the edges a little.