Cleaned up PlayerController, and added initialization skeleton
Took all the stuff out of PlayerController that we have decided that we are no longer using. Also added basic functionality for finding and opening an arbitrary config file from My Documents, as we will not have to ask for admin priviledges here. Added commented skeleton of a database connection and query, and added the ADDBPlayerList class to serialize the database data into when Team 4 has that part ready.