Roughly Implemented Config FileStream, and Encryption Skeleton
Implemented the opening of the the config file, and layed out where and how it will be decrypted, as we have yet to pick our RSA parameters. Created a utility class for encryption, and made ReinitializePlayer( ) that will be used in redressing the player everytime the local equipmentHash is changed.